icono

+++ Capa_Datos/Capa_Presentacion/Principal.cs
@@ -135,7 +135,7 @@ namespace Capa_Presentacion
                 if (xml.Consulta_Login(txtCedulaLogin.Text, txtContraseñaLogin.Text) == "Cliente")
                 {
 
-                    nombreUsuario = xml.Retorna_Nombre(Convert.ToInt16(txtCedulaLogin.Text));
+                    nombreUsuario = xml.Retorna_Nombre(Convert.ToInt32(txtCedulaLogin.Text));
                     this.Hide();
                     Parqueo conectar = new Parqueo();
                     conectar.Show();
